move to utils module
