+++ pypums/pypums.py
@@ -1,16 +1,12 @@
 """Main PyPUMS module."""
-from pathlib import Path
 
-from typing import Union
-
-from pypums.download import download_acs_data
-from pypums.utils import build_acs_url
+from pypums.utils import build_acs_url, download_acs_data
 
 
 def get_data(
-    year: Union[int, str] = "2018",
-    survey: Union[str, int] = "1-Year",
-    person_or_household: str = "person",
+    year: int = 2018,
+    survey: str = "1-Year",
+    sample_unit: str = "person",
     state: str = "California",
     download_path: str = "../data/raw/",
     extract: bool = True,
@@ -24,7 +20,7 @@ def get_data(
     URL = build_acs_url(
         year=year,
         survey=survey,
-        person_or_household=person_or_household,
+        sample_unit=sample_unit,
         state=state,
     )
 
@@ -37,18 +33,6 @@ def get_data(
     )
 
     return None
-
-
-def tree(directory):
-    """
-    Displays a directory's tree.
-    """
-    directory = Path(directory)
-    print(f"+ {directory}")
-    for path in sorted(directory.rglob("[!.]*")):
-        depth = len(path.relative_to(directory).parts)
-        spacer = "    " * depth
-        print(f"{spacer}+ {path.name}")
 
 
 if __name__ == "__main__":
